\docType{package}
\name{ecipex-package}
\alias{ecipex-package}
\title{The ecipex Package}
\description{
  Efficient calculation of fine structure isotope patterns
  via Fourier transforms of simplex-based elemental models.
}
\details{
  \tabular{ll}{ Package: \tab ecipex\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2014-08-04\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab true\cr }

  Provides a function that quickly computes the fine
  structure isotope patterns of a set of chemical formulas
  to a given degree of accuracy (up to the limit set by
  errors in floating point arithmetic). A data-set
  comprising the masses and isotopic abundances of
  individual elements is also provided.
}
\author{
  Andreas Ipsen
}
\keyword{package}

