\docType{data}
\name{nistiso}
\alias{nistiso}
\title{Masses, isotope patterns and nucleon numbers of the elements}
\description{
  A data frame giving the masses, standard isotopic
  abundances and nucleon numbers of most chemical elements
  as provided by the Physical Measurement Laboratory of the
  National Institute of Standards and Technology (but taken
  from separate publications). It includes the four
  variables \code{element} (the chemical symbol),
  \code{mass}, \code{abundance} and \code{nucleons},
  arranged so that each isotope is uniquely specified by
  one row.
}
\references{
  \url{https://www.nist.gov/pml/atomic-weights-and-isotopic-compositions-relative-atomic-masses}
}
\keyword{data}
\keyword{datasets}

