\docType{package}
\name{ecipex-package}
\alias{ecipex-package}
\title{Efficient Calculation of Fine Structure Isotope Patterns via Fourier Transforms of Simplex-Based Elemental Models}
\description{
  Provides a function that quickly computes the fine structure
  isotope patterns of a set of chemical formulas to a given degree of
  accuracy (up to the limit set by errors in floating point arithmetic). A
  data-set comprising the masses and isotopic abundances of individual
  elements is also provided and calculation of isotopic gross structures
  is also supported.
}
\details{
  \tabular{ll}{ Package: \tab ecipex\cr Type: \tab
  Package\cr Version: \tab 1.1\cr Date: \tab 2020-03-12\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab true\cr }
}
\author{
  Andreas Ipsen
}
\keyword{package}

